GSYE-757: Moved virtual heatpump tank energy parameters classes to their own module.
